+++ 0x04-more_functions_nested_loops/holberton.h
@@ -10,5 +10,5 @@ void print_diagonal(int n);
 void print_square(int size);
 void print_triangle(int size);
 void print_number(int n);
-long mypow(long base, long exponent);
+int mypow(int base, int exponent);
